% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcs.R
\name{arcs_track}
\alias{arcs_track}
\title{title}
\usage{
arcs_track(
  arcs,
  region,
  inverted = TRUE,
  x_axis = TRUE,
  arc_color = "score",
  color_function = grDevices::colorRampPalette(c("black", "red")),
  line_width = 0.5
)
}
\arguments{
\item{arcs, }{character | data.frame: Path to bed file with the arcs. Should have the following format: chr, start, end, score. Alternatively a data.frame can be supplied directly with the appropriate values. Column names are not considered. Thus, the columns should be in the right order.}

\item{region, }{character: string specifying region to plot.}

\item{inverted, }{bool: whether to invert the arcs.}

\item{x_axis, }{bool: whether to draw the x-axis.}

\item{arc_color, }{character: color of the arcs. If 'score' the arcs will be colored according to the score column.}

\item{color_function, }{function: color palette function used to map scores to colors.}

\item{line_width, }{numeric: width of the arc lines.}
}
\value{
A ggplot2 plot.
}
\description{
A short description...
}
