% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.R
\name{bed_track}
\alias{bed_track}
\title{Plot one or multiple bed tracks in same region.}
\usage{
bed_track(
  bed_paths,
  region,
  x_axis = TRUE,
  color = NULL,
  xlabel = NULL,
  legend = TRUE,
  linewidth = 2,
  ylabel = NULL,
  y_axis = TRUE
)
}
\arguments{
\item{bed_paths, }{named list | character: Either a path to a single bed file or a named list of bed files, where each entry in the list is the path to the file, and the name being the sample name (e.g., list('sample1' = 'path/to/sample1.bed', 'sample2' = 'path/to/sample2.bed').}

\item{region, }{character: string specifying region to plot (accepts chr:start-end and chr-start-end formats).}

\item{x_axis, }{bool: Whether to draw the x_axis.}

\item{xlabel, }{character | NULL: x-axis label.}

\item{legend, }{bool: Whether to draw color legend.}

\item{linewidth, }{numeric: Width of the lines in the bed plot.}

\item{ylabel, }{character | NULL: y-axis label.}

\item{y_axis, }{bool: Whether to draw the y_axis.}

\item{colors, }{NULL | vector | function: Either a character vector of colors to use (must be at least same length as number of files) or a color function determining how samples are colored. If NULL, will default to ggplot2 default palette.}
}
\value{
A ggplot 2 plot.
}
\description{
Plots a region over multiple bed tracks.
}
