% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwig.R
\name{bigwig_logo_track}
\alias{bigwig_logo_track}
\title{Plot attribution score bigwig logo track at specified chromosomal region}
\usage{
bigwig_logo_track(
  file,
  region,
  BSgenome,
  annotations_path = NULL,
  min_importance = 0,
  y_padding = 0.1,
  xlabel = NULL,
  ylabel = "Contribution",
  x_axis = TRUE,
  y_max = NULL
)
}
\arguments{
\item{file, }{character: path to BigWig file.}

\item{region, }{character: string specifying region to plot (accepts chr:start-end and chr-start-end formats).}

\item{BSgenome, }{BSgenome: A BSgenome object for the reference genome used. For instance pass BSgenome.Hsapiens.UCSC.hg38 for hg38 reference genome.}

\item{annotations_path, }{character | NULL: Path to annotation file. If provided, bars and names will be drawn under respective regions. Intended for finemo motif-search output, but will accept any tsv file with columns region_chr, region_start, region_end, hit_coefficient_global, and motif_name.}

\item{min_importance, }{numeric: Minimum importance-score cutoff for annotations. Any annotation under this threshold will not be drawn.}

\item{y_padding, }{numeric: Spacing between logo letters and annotations.}

\item{xlabel, }{character | NULL: x-axis label.}

\item{ylabel, }{character| NULL: y-axis label.}

\item{x_axis, }{bool: Whether to draw the x_axis.}

\item{y_max, }{numeric: Upper limit of the y-axis.}
}
\value{
A ggplot 2 plot.
}
\description{
Plots bigwig scores at chromosomal region as a logo track, where the height of each letter corresponds to the score at the respective coordinate. Intended for deep learning model attribution scores.
}
