% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwig.R
\name{bigwig_track}
\alias{bigwig_track}
\title{Plot bigwig track at specified chromosomal region}
\usage{
bigwig_track(
  file,
  region,
  color = "black",
  ylabel = "Read Count",
  xlabel = NULL,
  x_axis = TRUE,
  y_max = NULL
)
}
\arguments{
\item{file, }{character: path to BigWig file.}

\item{region, }{character: string specifying region to plot (accepts chr:start-end and chr-start-end formats).}

\item{color, }{character: fill color of plot.}

\item{ylabel, }{character| NULL: y-axis label.}

\item{xlabel, }{character | NULL: x-axis label.}

\item{x_axis, }{bool: Whether to draw the x-axis.}

\item{y_max, }{numeric: Upper limit of the y-axis.}
}
\value{
A ggplot 2 plot.
}
\description{
Plot specified region of bigwig file using ggplot2.
}
