% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{genes_track}
\alias{genes_track}
\title{Plot genes at specified chromosomal region}
\usage{
genes_track(
  region,
  txdb,
  orgdb,
  collapse = FALSE,
  fully_in_view = FALSE,
  x_axis = TRUE,
  color = "black"
)
}
\arguments{
\item{region}{character: string specifying region to plot (accepts chr:start-end and chr-start-end formats).}

\item{txdb}{TxDb object: Transcript database, e.g., TxDb.Hsapiens.UCSC.hg38.knownGene.}

\item{orgdb}{OrgDb object: Organism database for gene symbols, e.g., org.Hs.eg.db.}

\item{collapse}{bool: If TRUE, merges all transcripts of a gene into a single flattened structure.}

\item{fully_in_view}{bool: If TRUE, removes any genes that extend beyond the specified region.}

\item{x_axis}{bool: Whether to draw the x-axis.}

\item{color}{character: Color for the genes.}
}
\value{
A ggplot2 plot.
}
\description{
Plot genes and transcripts in a specified genomic region.
}
