% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_track.R
\name{install_pyGenomeTracks}
\alias{install_pyGenomeTracks}
\title{Install pyGenomeTracks Dependency}
\usage{
install_pyGenomeTracks()
}
\value{
None
}
\description{
Install pyGenomeTracks dependency for plot_gtracks()
}
\details{
The function will install miniconda if does not exits and
check pyGenomeTracks installation.
}
\examples{
\dontrun{
install_pyGenomeTracks()
}
}
\author{
Omar Elashkar
}
\keyword{plot_gtracks}
