% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwig.R
\name{multi_bigwig_track}
\alias{multi_bigwig_track}
\title{Plot multiple bigwig tracks in same region.}
\usage{
multi_bigwig_track(
  paths_list,
  region,
  colors = NULL,
  ylabel = "Read Count",
  xlabel = NULL,
  x_axis = TRUE,
  y_max = NULL,
  split_tracks = TRUE,
  legend = TRUE
)
}
\arguments{
\item{region, }{character: string specifying region to plot (accepts chr:start-end and chr-start-end formats).}

\item{colors, }{NULL | vector | function: Either a character vector of colors to use (must be at least same length as number of samples) or a color function determining how samples are colored. If NULL, will default to ggplot2 default palette.}

\item{ylabel, }{character| NULL: y-axis label.}

\item{xlabel, }{character | NULL: x-axis label.}

\item{x_axis, }{bool: Whether to draw the x_axis.}

\item{y_max, }{numeric: Upper limit of the y-axis.}

\item{split_tracks, }{bool: Whether to plot each bigwig file in a separate track.}

\item{legend, }{bool: Whether to draw color legend.}

\item{file, }{named list: A named list of bigwig files, where each entry in the list is the path to the file, and the name being the sample name (e.g., list('sample1' = 'path/to/sample1.bw', 'sample2' = 'path/to/sample2.bw').}
}
\value{
A ggplot 2 plot.
}
\description{
Plots a region over multiple bigwig tracks, either one-per-track or all stacked in one track.
}
