% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{set_theme_modern}
\alias{set_theme_modern}
\title{Set modern default ggplot theme.}
\usage{
set_theme_modern()
}
\description{
Applies a custom modern ggplot theme. This will unify themes across track plots, but will also apply to all ggplots in the R session. To reset the theme, run \code{set_theme()}.
}
