% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gtracks.R
\name{plot_gtracks}
\alias{plot_gtracks}
\title{Plotting genomic tracks}
\usage{
plot_gtracks(obj, ...)
}
\arguments{
\item{obj}{Object to be plotted.}

\item{...}{Other parameters to be passed.}
}
\value{
None
}
\description{
This is a generic function used to plot \code{genome_track} objects.
}
\author{
Omar Elashkar
}
