% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{track_hic_matrix}
\alias{track_hic_matrix}
\title{Generate HiC track}
\usage{
track_hic_matrix(
  file,
  title = NULL,
  height = NULL,
  overlay_previous = "no",
  orientation = NULL,
  max_value = NULL,
  min_value = NULL,
  transform = "no",
  rasterize = TRUE,
  colormap = "RdYlBu_r",
  depth = 1e+05,
  show_masked_bins = FALSE,
  scale_factor = 1
)
}
\arguments{
\item{file}{String. The location of the track file}

\item{title}{String. If specificed, the title of the track to be displayed.}

\item{height}{Numeric. The height of the plotted track in cm. Default is 2.
See notes.}

\item{overlay_previous}{String. Options are "no" (default) or "yes" or
"share-y".}

\item{orientation}{String. Set to "inverted" to make the track upside down.
Default is NULL.}

\item{max_value}{Numeric. Default is NULL. The max value cut-off for the
numeric column.}

\item{min_value}{Numeric. Default is NULL. The max value cut-off for the
numeric column.}

\item{transform}{String to indicate type of transformation applied.
Default is "no".}

\item{rasterize}{Boolean. Default is FALSE.}

\item{colormap}{String with matplotlib-compatible colormap. Default is set to "viridis".}

\item{depth}{Numeric value above 1 to indicate the maximum distance that should be plotted. Default is 100000.}

\item{show_masked_bins}{Boolean. If TRUE, showing masked bins as white lines. Default is FALSE.}

\item{scale_factor}{Numeric factor by which matrix is to be scaled.}
}
\value{
genom_track
}
\description{
Create a genome_track for matrix files. Currently, only cool format and h5 format.
}
\details{
This function expect cool or h5 format. Format converter like
\href{https://hicexplorer.readthedocs.io/en/latest/content/tools/hicConvertFormat.html#hicconvertformat}{hicConvertFormat}
can help converting to supported formats.
depth is the maximum distance that should be plotted.
If it is more than 125\% of the plotted region, it will be adjsted to this maximum value.
colormap argument should be compatible with \href{https://matplotlib.org/2.0.2/users/colormaps.html}{matplotlib}.
show_masked_bins plots bins not used during the corrections as white lines.
Setting this argument to FALSE (default) extends neighboring bins to
obtain an aesthetically pleasant output.
scale argument scales the matrix by specific factor.
This is useful if plotting multiple hic-matrices to be on the same scale.
}
\author{
Omar Elashkar
}
